Pysetter/requirements.txt
opencv_python
matplotlib
torchsummary
imageio
python-docx
django
